% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rgff.R
\name{check_gff3}
\alias{check_gff3}
\title{Test consistency and order of a GFF3 file}
\usage{
check_gff3(gffFile)
}
\arguments{
\item{gffFile}{Path to the input GFF3 file}
}
\value{
A data frame with the errors detected
}
\description{
This function test consistency and order of a GFF3 file.
}
\keyword{internal}
