% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rgff.R
\name{get_pairs_from_gff}
\alias{get_pairs_from_gff}
\title{Creates a features pair file from a GFF file}
\usage{
get_pairs_from_gff(
  inFile,
  outFile,
  fileType = c("AUTO", "GFF3", "GTF"),
  forceOverwrite = FALSE
)
}
\arguments{
\item{inFile}{Path to the input GFF file}

\item{outFile}{Path to the output features pair file, if not provided the output will be inFile.pairs}

\item{fileType}{Version of the input file (GTF/GFF3). Default AUTO: determined from the file name.}

\item{forceOverwrite}{If output file exists, overwrite the existing file. (default FALSE)}
}
\value{
Path to the generated features pair file
}
\description{
This function creates a features pair file from a GFF file
}
\keyword{internal}
