% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rgff.R
\name{get_pairs_from_gff3}
\alias{get_pairs_from_gff3}
\title{Creates a features pair file from a GFF file}
\usage{
get_pairs_from_gff3(gffFile, outFile, forceOverwrite = FALSE)
}
\arguments{
\item{gffFile}{Path to the input GFF file}

\item{outFile}{Path to the output features pair file, if not provided the output will be gffFile.pairs}

\item{forceOverwrite}{If output file exists, overwrite the existing file. (default FALSE)}
}
\value{
Path to the generated features pair file
}
\description{
This function creates a features pair file from a GFF file
}
\keyword{internal}
