% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rgff.R
\name{get_pairs_from_gtf}
\alias{get_pairs_from_gtf}
\title{Creates a features pair file from a GTF file}
\usage{
get_pairs_from_gtf(gtfFile, outFile, forceOverwrite = FALSE)
}
\arguments{
\item{gtfFile}{Path to the input GTF file}

\item{outFile}{Path to the output features pair file, if not provided the output will be gtfFile.pairs}

\item{forceOverwrite}{If output file exists, overwrite the existing file. (default FALSE)}
}
\value{
Path to the generated features pair file
}
\description{
This function creates a features pair file from a GTF file
}
\keyword{internal}
