% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rgff.R
\name{gff3_to_paths}
\alias{gff3_to_paths}
\title{Creates a features path file from a GFF file}
\usage{
gff3_to_paths(gffFile, outFile, forceOverwrite = FALSE)
}
\arguments{
\item{gffFile}{Path to the input GFF file}

\item{outFile}{Path to the output features path file, if not provided the output will be gffFile.paths}

\item{forceOverwrite}{If output file exists, overwrite the existing file. (default FALSE)}
}
\value{
Path to the generated features path file
}
\description{
This function creates a features path file from a GFF file
}
\keyword{internal}
