% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rgff.R
\name{gtf_to_paths}
\alias{gtf_to_paths}
\title{Creates a features path file from a GTF file}
\usage{
gtf_to_paths(gtfFile, outFile, forceOverwrite = FALSE)
}
\arguments{
\item{gtfFile}{Path to the input GTF file}

\item{outFile}{Path to the output features path file, if not provided the output will be gtfFile.paths}

\item{forceOverwrite}{If output file exists, overwrite the existing file. (default FALSE)}
}
\value{
Path to the generated features path file
}
\description{
This function creates a features path file from a GTF file
}
\keyword{internal}
