% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rgff.R
\name{saf_from_paths}
\alias{saf_from_paths}
\title{Generates SAF formatted data from a features path file for the given blocks and features}
\usage{
saf_from_paths(
  pathsFile,
  groupBy = c("mRNA", "gene"),
  block = c("exon", "CDS")
)
}
\arguments{
\item{pathsFile}{Path to the input features path file}

\item{groupBy}{Vector of features to group by in feature count}

\item{block}{Vector of features to be used as block in feature count}
}
\value{
SAF formatted data frame for the given blocks and features
}
\description{
This function creates a SAF file from a features path file for the given blocks
and features
}
\keyword{internal}
