% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rgff.R
\name{sort_gff3}
\alias{sort_gff3}
\title{Sorts a GFF3 file}
\usage{
sort_gff3(gffFile, outFile, forceOverwrite = FALSE)
}
\arguments{
\item{gffFile}{Path to the input GFF3 file}

\item{outFile}{Path to the output GFF3 file, if not provided the output will be  the input path (without extension) with the suffix sorted.gff3}

\item{forceOverwrite}{If output file exists, overwrite the existing file. (default FALSE)}
}
\value{
Path to the sorted GFF3 file
}
\description{
This function produces a sorted GFF3 file from an unsorted GFF3 file.
The output file will be sorted by Chromosome, Start, End (reverse) and feature (based on the precedency in feature tree)
}
\keyword{internal}
